[man] : Failed to filter valid parameters: The term 'Filter-ValidParameters' is not recognized as the name of a cmdlet, function, script file, or operable program. Check the spelling of the name, or if a path was included, verify that the path is correct and try again. At line:1 char:1 + . 'G:\OperaLauncher\man.ps1' + ~~~~~~~~~~~~~~~~~~~~~~~~~~~~     + CategoryInfo          : NotSpecified: (:) [Write-Error], WriteErrorException     + FullyQualifiedErrorId : Microsoft.PowerShell.Commands.WriteErrorException,man.ps1
G:\OperaLauncher\man.ps1 : Process block failed: The term 'Filter-ValidParameters' is not recognized as the name of a cmdlet, function, script file, or operable program. Check the spelling of the name, or if a
path was included, verify that the path is correct and try again.
At line:1 char:1
+ . 'G:\OperaLauncher\man.ps1'
+ ~~~~~~~~~~~~~~~~~~~~~~~~~~~~
    + CategoryInfo          : NotSpecified: (:) [Write-Error], WriteErrorException
    + FullyQualifiedErrorId : Microsoft.PowerShell.Commands.WriteErrorException,man.ps1

Filter-ValidParameters : The term 'Filter-ValidParameters' is not recognized as the name of a cmdlet, function, script file, or operable program. Check the spelling of the name, or if a path was included, verify
that the path is correct and try again.
At G:\OperaLauncher\man.ps1:254 char:29
+             $filteredHash = Filter-ValidParameters "Clear-Cache" $cac ...
+                             ~~~~~~~~~~~~~~~~~~~~~~
    + CategoryInfo          : ObjectNotFound: (Filter-ValidParameters:String) [], ParentContainsErrorRecordException
    + FullyQualifiedErrorId : CommandNotFoundException